More clean up. Moved merkle tables into `fuel-core`, because it is internal details of the `Database`. Fixed typo "teh" Implemented remove.
