+++ crates/types/src/entities.rs
@@ -1,6 +1,5 @@
 //! Higher level domain types
 
 pub mod coin;
-pub mod merkle;
 pub mod message;
 pub mod resource;
